requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.11.0
segmentation-models-pytorch>=0.3.3
pyyaml>=6.0
requests>=2.31.0